backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
requests>=2.26.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.0.0